# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasks API",
        "version": "1"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authorization",
                "description": "Authorization and return API-token",
                "operationId": "72bf786ad8ff53c7391309c836350a8d",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API-token"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List of tasks",
                "description": "List of tasks by filters",
                "operationId": "19db9bcf6cfc82a01437fdb46b235ed7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Task status for filters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Priority of task for filters",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query string for filters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query string for filters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Id of user for filters",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort1Field",
                        "in": "query",
                        "description": "Field name for first sorting (priority,completed_at,created_at)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "priority"
                        }
                    },
                    {
                        "name": "sort1Dir",
                        "in": "query",
                        "description": "Direction of sorting of first sort field (asc,desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "sort2Field",
                        "in": "query",
                        "description": "Field name for second sorting (priority,completed_at,created_at)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort2Dir",
                        "in": "query",
                        "description": "Direction of sorting of second sort field (asc,desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The list of tasks in JSON"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Add new task",
                "description": "Add new task",
                "operationId": "dd7609a4e87fc90c63f0b8a4608d8335",
                "parameters": [
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Task priority",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created task in JSON"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Tree of one task",
                "description": "Tree of one task",
                "operationId": "ed7b7e2db52b8c3733cecd26db2e7f65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tree of on task in JSON"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "description": "Update task",
                "operationId": "e03c2831dc2e9b28be4930d51f76c94f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Task status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Priority of task",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of task",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of task",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tree of on task in JSON"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "acc43a5a4bc196d969b047c7a45e7cdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok if success, error message if error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{id}/done": {
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Set task as done",
                "description": "Set task as done",
                "operationId": "3b602aa73f7384d1b3c3a2a1f088e2f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok if success, error message if error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}